Draw now takes Rect, change from shared to *, set returns this...
Changed over to using a Rect for position and size when drawing.
Previously we only used x and y ints, that had to change.
Moving away from shared_ptrs in favour of normal pointers.
Mostly because I want to learn more about managing memory and lifetimes,
but secondarily even now, dealing with them is kind of painful.
Some methods (mostly setters) will return the pointer to the object,
to aid method chaining. Especially when creating Widgets.
Lastly changed colour of test widgets to be easier to see and distinguish.