Got MouseMotion events working...
To facilitate it we gave every widget a unique id,
though the way we accomplished it could be improved.
Anyway this makes it easy for us to tell when a mouse
entered or left a widget's bounds and we just fire the appropriate event.
Regardless we fire the motion event.

For the widgets to support things like hover etc.
I will need to add a state enum.
Then its easy to add a change of color when a widget is hovered .
Just doing it when the mouse is over the widget does
not really work since the renderer gets all glitchy.

Also the widgets will probably need their own default callbacks for things like hover.