Removed Application* from Button/Box constructor...
This provides a nicer API but its pretty strict when it comes to structring your app.
Next commit I will focus on improving the ergnomics of this.
Current idea is to have every widget specify its parent as the first argument.
The upsides of this would be that we dont have to worry about
the order of the append call, moreover the user would not need to use append.