Improved app/widget API to be more relaxed when building the widget tree...
Now setting the main widget or appending widgets should not affect the app.
This is accomplished by passing some required data when calling run, and
by using append at run time to do the same.
The downside of this approach is that it adds boilerplate that is required
when implementing a new widget. However that burden can be eased later
on with a macro, since the code should be the same for every widget.