Eliminated the need for id...
I seem to recall trying this once before and it not working...
I dont really know whats different, it might have been
 the logic around it rather than the comparison.
Anyway it means that now we dont need void*app nor do we need id.
So all that effort from previous commits is somewhat wasted.
But at least we get a nice api with reduced complexity,
furthermore we should be able to remove dc* since we pass it anyway on draw.
Though size_hint() might need to take a dc argument for measuring text and such.