Split app/widget into cpp/hpp files, fixed segault when handling mouse up...
This finally allows us to have a pointer to app which
is needed for issuing updates.
Which means we can keep our current API without
specifically passing in a parent or the app itself.
Next up I will work on remove the need to call
set_main_widget manually.
Ideally this will just be handled when creating application by itself.
The current idea is to simply create a new VBOX which takes up the entire window space.
Add an append function to app which would be equivalent
to calling append on the box itself and bobs your uncle.