Fixed bug which required a call to show twice...
This was caused by premature optimization.
I wanted to not have to compute the dimensions of text more than necessary.
So the size_hint for button would only update when drawing.
As a result after a second draw call was required if the text changed.
To change it i added pointer to dc for size_hint() so that
we can compute various things when the layout needs to know the extent of the widget.