Improved fill on widgets...
Previously there was a problem where if the last widget was larger than others
every widget before it would start expanding assuming that the last widget has the same size.
Now this behaviour has been fixed, simply get the min required
space for all the widgets and subtract it from the
available expandable space, finally divide that between
all widgets equally (but in the future we can specify the proportion so that
a widget can expand twice as much as the others).