scrollbar: Size and scrolling improvements...
Now length of the Slider reflects the visible area
 proportionally to the total scrollable area.
 This also leaves the Slider more customisable than before.
 Scrolling with the mouse has been fixed.
 Now instead of using the mouse position we use the
 relative movement of the mouse to change the value.
 This prevents snapping of the Slider when starting to drag
 from somewhere other than the very top.
 This would happen because the new value would change
 to the current mouse position. Now the old value is changed
 by the relative movement of the mouse.
 Which means that you can now click anywhere on the
  slider and get the same scrolling experience.
Finally the movement of the Slider is fully accurate.
Meaning that the moment the end of the Slider reaches
the end of its region is the moment when the value hits 1.0 ie max.
Previously you still had to move the mouse to the very bottom of the Slider.
As a result the Slider would stop moving but the value wasn't maxed out yet.