Improved text rendering performance...
We pass the app* to the dc which can then pass it to renderers.
Though actually seems we could just pass the size instead.
Anyway this means that we can reason about which text characters to draw based on the visiblity.
So for a really long line of text, say 2000 chars we only draw what is visible on screen currently.