Change click callback to std::function...
Eventually all callbacks will use std::function.
The advantage of using it over a normal function pointer
is that we can bind a member function a lambda,
or any freestanding function that matches the parameters.
This simplifies the process of binding callbacks
since you now get the context around the function.
This makes it much nicer to access the necessary data,
like anything from the class that houses the callback.