Add focus members, change callback from click to down...
Click produce unwanted behaviour
because on mouse down the button would be pressed
which would cause the ui to update
drawing the insert cursor at the old location
and only after the click callback would it update again to the new location.
Furthermore i can see how mouse down would be better
in the future once we support text selection using the mouse.