Add SDL_TextInput and add text to focused element...
This basic implementation is very naive.
The biggest problem is that it casts to Button*.
The other problem is that is appends the incoming text to the widgets text.
In the future the text should be passed to the widget
and handled by the widget from there.
Secondly there should be an abstract text input class instead of casting to button.
^ actually there probably needs to be an abstract accepts keyboard input
class and then another text input class or something along those lines.