Added NewWidget, this is a direct replacement for Widget...
For the moment its called NewWidget to prevent name clashing but
in the future this will be renamed to Widget.
Anyway started following a coding style.
Adding documention comments as I go.
Also trying to better decisions about the class composition.
For example this time around this will be an abstract class
so youre forced to implement the necessary methods.