Force a size recalculation for layout when size_changed...
This is necessary because we dont calculate the total_widget_size anymore.
Which is because that is the same as sizeHint and
in the future we will want to eliminate the loop that that calculation was a part of.
The sizeHint is never called for the mainWidget
because its the layout that calls sizeHint on other widgets :^)