Major reworks to fonts and textures...
Added a font struct for representing different fonts
these could be passed to text drawing and measuring functions
however for the moment we only have the default_font which is part of DC.

In order to facilitate more than one active texture I wrote
an implementation that queries the amount of available slots on the hardware
and uses all of the in the drawing process binding and unbinding textures as needed.
In order to facilitate the above the fragment shader for text (and later on the pipeline)
has to be dynamically generated at runtime since we need
the amount of texture slots queried in opengl and
we need to generate the branching code which is used
to select the correct texture from the ones currently activated.

Finally updated shader. Added an empty constructor.
It is used because we want to add a shader at runtime
and id rather not deal with extra pointers.
Temporarily removed the version which reads from file.
I mostly removed it because of exceptions. At some point ill
probably rewrite to into a c style file read.