Pass font parameter to text funtions, add font to widget...
Basically what it says on the tin.
Widget now gets a font member.
In Button you can see the implementation where
it will use the font member if its not null
otherwise it will use the default font from DC.
Finally in main we made is so that every button in the middle layout
will use the other non-default font.