Added funtion to drawImage at a specified size...
This would be particularly helpful in the future
when many widgets share the same texture but some want to draw it
at a different size.
Though i can see this being problematic since the texture might not have
enough space because the sizeHint wouldnt take that size into account
but rather the original or whatever is set.