Updated button...
Im not entirely happy with it mainly due
to the face that the alignment of the image and text
isnt quite right.
This is caused by the the text x, y getting normalized.
So the text is never quite centered.
The plan for now is to change to using either int rects,
or to use int values and only changed them into floats when passing them to openGL.
This should fix any issues with text and it should not need to be normalized.
The biggest things we will need to change is the space allocation
in box layout. Though there might be other things that need to change.