Change the way that widgets expand in the opposite direction of the layout...
Previously they would take up at maximum the rect size
which for a small application is often the window size itself.
The change makes it so that they will now expand to the maximum size of the layout.
So in a horizontal layout of three widgets: 100, 300, 200px tall,
each of them will expand upto 300px since thats the max in that layout.
To be honest im not 100% sure this is desired behaviour
but until i build something concrete we wont really know.