Fixed bug with weird content layout after removing a scrollbar...
The problem was the scrollbar wasnt being removed :^)
It was still being drawn but You couldnt see it because it was outside
of the given rect, which made it a bit more difficult to figrue out.
It was caused by me removing the call to remove the scrollbar
before figuring out how to add both of the scrollbars correctly
and later not putting it back in.