Fixed an uninteractable scrollbar in scrolledbox...
The issue would manifest itself when the addition of a horizontal scrollbar
would cause a vertical scrollbar to show up.
The scrollbar was a actually interactable but that would never show
because it would be added and removed every update.
The fix is to only remove the vertical scrollbar at the end.