Use sdl swap interval to limit framerate...
We need to manually keep track of the mouse movements
since otherwise the movement would not be accurate.
Otherwise the movement lags behind because mouse updates
more than the vsync rate.
And alternatively when we flush the mouse events it works fine
on windows but on linux the relative mouse movement isnt enough
to keep up wit the mouse.