Removed scissor from box...
This is in preparation for implementing the clipping using the fragment shader.
The clipping ATM doesnt work well for clipped widgets within clipped widgets
because the clipping rectangle gets overwritten
and then disabled.
But all of that should be fixed after implementing our own clipping.
The idea is that you can set a clipping rect in the drawing context
and each subsequent drawing command will use that clip rect until it is disabled.