Update NoteBook to use a TabBar...
The NoteBook itself is still quite WIP but this helps
to bring it along. Makes for much easier to reason about code.
Since the tab buttons are actual widgets now and they
get managed by the tab bar.
Events still dont work because NoteBook is not a layout.
But in this case the plan is to just override the propagte mouse event function.
Instead of checking the children we will just
check against the tab bar and or against the current tab
and proceed as normal from there.