Change propagating mouse events for layouts...
Previously if the mouse was over a layout but not
any of its widgets this function would return a nullptr.
Now if the event doesnt match any of the layouts children
it is instead sent to the layout itself.
This is needed for the addition of mouse wheel scroll events.