Change from m_cursor to selection.begin...
Previously the m_cursor vars were just tagging along with the rest
but they werent needed for anything that we couldnt use selection for.
In the future we probably will offer a getCursorIndex() function
which can just return the selection index.

When drawing the insertion cursor we now use selection.end.

Note that currently the keyboard nagivation works but it has selection.
In the next commit we will include a check in the keyboard nav functions
to see if shift is pressed so when it isnt
we can set selection.begin to end after a nav event like moveCursorLeft().