Vastly improved the test runner...
The tests to run are now collected automatically from the tests directory.
The test runner now relies on make to build and keep the object files for the lib up to date.
Object files for the lib are automatically collected from the build dir.
The above makes the current build command much quicker
because we only build the test file and link it against objects.
Finally when a compilation of a test fails it is factored into the test stats at the end.