Implement capturing the mouse...
The mouse is captured on every mouse down event
and released on every mouse up event.
When the release happens outside the window the event is not sent to the window.
This finally allows use the scrollbars outside of the window bounds
and previously this could get annoying because suddenly the scrolling would stop
since the mouse left the window.