Basic skeleton of the model (super WIP)...
Im not really happy with it atm.
ATM this is just a port of the API that I used in the typescript version.
I will need to rethink this.
Its not very nice to use.
The append method on Tree is also full of compromises where
i just wanted to get it working in c++.
Task for tomorrow is to try and design a nice API thats not as error prone.