Update tree model implementation...
Change yesterdays implementation to be more idiomatic.
Instead of using TreeIter, which was proving very cumbersome,
we simply use pointers instead.
This should be straight up faster (since we only need to deref a pointer
as opposed to recursive vector element accesses)
but also nicer to work with.
Removed is visible since that seemed redundant with is collapsed,
because when something is collapsed it is not visible and same goes for its children.
Added destructor code which should properly delete the entire tree.
Changed hidden to T* so that we can always call delete on it.
Added a clear method for getting rid of all the data within the tree.
Updated main example to see if the tree works.