Change behaviour of collapse function...
Previously when a node was already collapsed
the collapse function would expand it.
This is the behaviour that we want when clicking on a node
but its undesired behaviour when calling the collapse function.
The function does something different than what the name suggests.
Changed collapse function to only collapse but keep the same behaviour when clicking a node.