Changed behaviour of layout()...
Previously layout would globally signal the app that.. well.. the layout has changed.
This was used by some widgets, mainly the ones that had children that could change size.
The original reason for having it was: because back in the day
we didnt have a parent member so the widgets with children couldnt tell
that they needed to update their layout when their children changed.
Which led to this mess of each widget both: having a
m_size_changed member and needing to tell the app that layout has changed.
So everytime something would cause a layout shift
we needed to do three things: call update, set m_size_changed and call layout.
Now the new behaviour of recursively telling the parent that the size has changed
eliminates the need for manually setting the member and
changes layout to do that automatically.
Furthermore, layout now calls update itself since there was never
a reason not to call update if the layout did change.
So now if doing something changes the size of a widget
you just need to call layout() and nothing else.
Oh and this is more efficient since when the layout change
only affect one part of the window the other aren't recalculated.