Added a table mode...
This affects the way in which the data is drawn.
Mainly there is no  collapse / expand button.
There is no indentation.
The data looks like a table but it isnt.
It will still only draw those nodes that are not collapsed.
Furthermore i decided to allow the user to expand and collapse
nodes programmatically just not with the mouse,
so keep that in mind!