Simplify texture struct...
In the future texture and image are not going to be interchangable.
This is for a few reasons.
The biggest one is that we want the option to have shared textures.
This allows us to reuse textures and display them at
different size, colors, flipped etc.
without affecting the original.