Rework of image...
This new iteration has a shared ptr at heart.
So we can use the same texture across multiple images etc.
But you can still make an image by directly creating a texture for it
with either a path or some memory parameters.
Furthermore this new image can have its size manually changed.
This can be reverted by querying the original size of the texture and setting the min size.
Furthermore some stuff like flipping the image
have been moved from texture to the image class.
And finally we now have the option to actually change the texture of the image.
We can also return the current texture so that other images can use it.