Changed the way Column works...
Moved adding the sort icon into the Column ctor.
It just makes more sense for it to be there.
Now the column has an internal button with no border.
This can have an image and a label.
The horizontal alignment of it is controlled with
internal spacers as needed.
Slightly change default min width.
Simplified the sizeHint calculation.
The main difference is that we always take hidden widgets into account.
This helps to set the best size for the column when setting a TreeView model.