Change over to non pointer style...
This is better because each widget has default style values.
And by default those simply use the default style.
We dont need to check if a style exists so we can always simply change its various attributes.
Furthermore we dont need keep track of its lifecycle and it should be cheap enough to copy.