Breakup some tests into examples...
This makes more sense because some of the tests
were really meant as examples to begin with.
They were in the tests folder so we could run them when running tests to make sure things are working.
Now there is a clear separation with examples
being used to show off how to do or use X thing.