Added a pretty cool treeview example...
Its standalone (at least on linux) so you
can just build and run it without the rest of the source.
Do note that you have to cd into the examples/treeview folder and then call make.
We still need to test it on windows.. we'll probably
just need an sqlite dll or something.