Re-added border/margin, fix one visual bug...
Previously i removed the border / margin from treeview
because there was a graphical bug that made the columns
get shown from underneath the scrollbar.
This bug has now been fixed, and was caused by incorrect clipping.
This happened because the column draw method
would override the clip set by the treeview
by setting it to the rect it got to draw into.
This rect could overlap the scrollbar, but this has since been fixed.
However there is still one more graphical bug which
manifests itself with some whitespace (color dependant on the fillRect of treeview)
on the left of the vertical scrollbar if the horizontal scrollbar is not maxed out.