Start on drawing treelines between nodes...
The very first step is operational. Namely:
We correctly identify the first root node for the visible nodes.
Then we descend it to then end and draw a line
for all its top children thats the correct length.
There are some small adjustments to make.
Like centering the beginning and the end of the line in the middle of their respective nodes.
Option to draw it or not, change color etc.
The next step is to recursively descend the root nodes children
and draw the lines to them as well.