Fix incorrect handling of mouse events...
Previously the x and y would not accurately correlate
to tree nodes, the disparity would increase the more we scrolled.
This was happening because the values we were using
did not account for the size of the scrollbars,
and so overtime the x and y would become inaccurate up to the size of their respective scrollbars.