Fixed a problem with mouse up events not propagating...
This was caused by a copy paste error a long time ago.
Because of it the comparison for the y axis was wrong
and so the events sometimes would not propagte when they should have.
Also small rant: I was trying to use the vs code built in debugger (gdb)
to get the the heart of this bug
but about 50% of the time hitting a breakpoint would cause my entire desktop environment to freeze.
Why? Good fucking question!