Add tooltip to widget...
Added code necessary for displaying tooltips within a graphics context.
Note that these tooltips can only be shown within
the window, and NOT over or outside.
ATM they are very simple and lack customizability.
Next commit I will probably add checks to reduce the
possibility of the tooltip not being visible to the user.