Better tooltip positioning...
Simply check if either the width or the height
of the tooltip will make the tooltip get drawn
outside of the window's extens, and
change the position its drawn at to be the opposite side.
When the window is too small this can result
in the tooltip not being visible but this
should only happen at pretty small window sizes.