Removed headless to simplify building / testing process...
The headless solution had problems because it required us
to build all the binaries with a -DTEST flag.
That was bad because it meant we couldnt run that binary
normally. Instead we would have to rebuild without it to run the program normally without quitting.
The new solution uses commandline arguments instead.
This way there is only one version of the binary
and if we want to quit immediately for running the tests
simply pass quit as the first cmdline argument.

The next step is to simplify the build process for tests and examples.