Changed the color implementation...
Add extra options to construct the color from uint8s or a string.
Added an extra member to denote whether the color
should defer to the default in the dc default style.
Update all instances of color throughout the project
to use the 1.0f float syntax and changed some
instances to macro defines.