Add EmptyCell when a cell has no visual value...
This is what i'm using currently for the mhw db test app, but
im not so sure this is the right move because that would mean
we would need to be very careful accessing column contents.
Perhaps in the future I will change is so that every cell renderer class can have an empty
ctor.