Overhaul of tree traversal, speedup...
Primo I changed the bool defines to an enum class since
I needed at least try states so bool couldnt work anymore.
The three states now are:
Continue - Keep going as normal and traverse all nodes of the tree
Next - Prematurely end the traversal of the current node and switch to the next available one
Break - End the traversal asap
Note: That break usually still needs to get to the end of the current root node
but usually that should happen fairly quickly but its perhaps not optimal.

Furthermore I added a function to handle the early exits automatically.
Now there is the descend() which requires manual handling and
forEachNode() which is automatic but less flexible.

Last but not least this results in a pretty major speedup on large datasets.
Previously ending the traversal early did not account for the case
where the treeview was composed of only root nodes.
So in a table with 100,000 rows even at the very top
the cpu would still go over every single node and cell
when it should have stopped at the edge of the visible ones.