No longer take the height of the scrollbar when calculating sizeHint...
This was problematic because the addition and subtraction of the scrollbar caused
the size to change and would require another layout and update pass.
This pass would need to follow immediately after the last update because it
was already to late to rectify the allocated rect.
Furthermore I didnt want to always add the size of the scrollbar:
a) because it was proving somewhat problematic without always having an instance of the object
b) because without the scrollbar the tab buttons looked odly big in comparison to everything else.

The current solution is to have a tiny scrollbar that fits into the default bottom padding.