Rewrite of slider...
The feel of the slider has been greatly improved.
Instead of just using relative mouse movement
we use a combination of mouse position and mouse movement, and
we switch between the two based on where the mouse is.
The slider value now comes from the middle of the slider button, and
because of it the slider behaves very similiar to what people are generally used
to (like vscode for example).
The only slight issue is that when dragging the slider on the far end on either side
when the mouse gets to the min / max values the slider will snap to that value a little.