Finally add uniform scrolling, atm only for Scrollable...
This was a long time coming but finally
the scroll wheel scrolling behaves as you would expect.
It always scrolls by the same amount of pixels (atm defined in the app).
Right now we do this by utilizing the step member of the slider.
The Scrollable base widget simply adjusts the step value based on the total virtual size.
For the moment only Widgets that inherit from Scrollable benefit from this.
I already knew I would need to move over some stuff like the ScrolledBox
over to Scrollable but this just
means its going to happen sooner rather than later :).