Fixed glitchy text (hopefully), cleanup...
I thought the previous solution that reused the old texture slot for text was good.
But i guess opengl is a fricking liar!
Because the texture slot should still be active!
The font atlas texture should already be bound!
And they are.. except when i use them for drawing it comes out scuffed?
but not always??
 docs.gl said that the texture should still be good to go unless i bound another one,
 which i didnt, but i guess you cant trust them because
 they say my version of gl should have at least 48 texture slots
 but it only has 32?
 IDK, this current solution seems to work?!?!?!?!??!