Add a clipTo function for rect...
This is incredibly usefull for correct clipping.
Essentially it allows you to transform a rect so that
it gets cut down to some parent rect if its too large,
or leaves it alone if it fits within the parent rect,
or if its completely out of the parents bounds it limits its size to avoid unwanted drawing.