Fix clipping issues involving box...
The issue was that the box always gives each
widget the size that they request even if the box
itself is given a smaller area.
This results in clipping issues when the box contains
other widgets which perform clipping.
In this scenario they then clip to their given rect
which is always what they request and this results
in some widgets drawing outside of their allocated areas.
The solution is to give each widget the size they request
for as long as the box has that much to give.
Otherwise give them whats left.
This leaves us with one issue which shouldnt be too difficult to fix.
Namely, things that rely on centering like text and images
draw outside of their area if they dont have enough space to center.
The solution is simple, if they dont have enough space to center, align them to the left | top.