Add a new Unroll mode to treeview...
This unroll mode changes the treeview from a scrollable to
a list that requests its actual size and does not scroll
unless its parent scrolls its contents.
This is useful for when you want treeviews
as part of some window/container that can have multiple scrollables.
This way the scroll event simply scrolls the parent's view
and everything is still visible since the tv basically behave like a box.