Remove size limiter...
The reason it was added initially is because
it made it easier to deal with deeply nested
widget hierarchies.
However, it doesnt make sense for a widget
with a virtual area and viewport to artificially limit its children.
Furthermore, because of the limit we had some uncanny looking widgets/layouts.
For example scrollbar sliders would resize based on how much of the widget you could see
rather than the actual size of the widget and its contents.