Fix and update select...
Change to size_t as int doesnt make sense for an index.
When changing the traversal code I left out the neccessary check for exiting early.
As a result once the index was found `i` was never incremented
which meant that every subsequent node was selected.
Changed over to the simpler forEachNode function.