Removed box size limiter as it was no longer needed...
Previously this was necessary to draw complex hierarchies of deeply
nested widgets.
After running the clipping tests and testing on mhwi_db it seems
that this is no longer true.
This is great but it is slightly worrisome as I was unable to track
down the exact changes that made this unnecessary.