Removed drawing_rect as it was no longer needed...
See the previous commit about boxes giving out space.
Now we can finally guarantee again that a Box or a ScrolledBox will
always give the requested size to a Widget.
This is because as long as there is at least one ScrolledBox at the
top of the widget tree every widget will have enough space.
Otherwise if there is a Box at the top instead then the contents will
go outside of the window anyway so this will not be a problem.