Remove scale option for text...
I actually did get it working but there were a couple of problems.
1. The filtering would need to be altered depending on the scale
factor. Scaling larger than 1.0 looks better with GL_NEAREST while
scaling smaller than 1.0 looks better with GL_LINEAR. Scaling 1.0 looks
good regardless. So there is an issue of changing the filtering method
based on the scaling factor, dynamically, whenever we draw text.
This would potentially need to be called on a per fillText call basis
as they could have different scales passed in.
2. Perhaps the bigger issue was that I was unable to get the correct
font information in regards to vertical space. Something about the
scaling made it so that on scaling over 1.0 there would be an issue
with some letters going outside of their allocated are at the top.
However because it seems that in that scenario there would be some extra
space at the bottom, that this issue is likely fixable.
I would just have to figure out what is creating that offset.

I'm open to re-adding this in the future, its just not a priority ATM.