Fix resizing treeview column behaviour...
It was mostly workable, however anytime you would change the size of
a column on the right side of a column that automatically expands, the
speed and at which it would change size proved problematic to control.
But it worked fine for the expandable column itself or for any column
to the left of it.

I changed the resizing to use the relative motion of the mouse but with
the exception that if the mouse is moving right and is outside of the
column then we ignore those events. This prevents the issue where
the mouse position doesn't really match the size of the column.