Made events works for Widgets draw within a TreeView...
This involves a revamp of how the event handling works for the tv
itself. Before we used to check which node we hit in the different
event callbacks, but now we do that step before in propagateMouseEvent.

Now anytime an event comes into the tv we check the usual suspects and
then we go down the Node tree to find a Widget or a Node that matches
the event position.
If no Widget matches we keep track of the Node and pass the event
down to the tv as normal. Afterwards you can simply access the Node
in the event callbacks using m_event_node.