Improve ColorPicker picking behavior by using an image...
Previously the ColorPicker would return the color of the cursor when
the mouse position happened to overlap it (since there is always some
delay between the mouse and the drawing).
By using an image instead of a dynamic gradient we can get rid off it.
Now we can just query the color from image pixel data which is not
affected by the drawing operations.