Fix reading color from texture, position is always relative...
Fixed bug where the color reading from the gradient texture was
incorrect, see comment.
Forgo continuously asking for the texture data and temporarily storing
it on the stack. Instead only retrieve it once and store it on the heap.
From now on the m_position Point is always relative to Widget.
So the upper left corner is 0, 0 rather than the actual position in
window space.