Bring back frame limiter for hopefully the final time :)...
This approach seems to be a good compromise between no fps cap and
a forced vsync through SDL2.

The downside of the SDL2 vsync is that it messes with events too much.
Its difficult for events to be kept up to date and we would need
several instances of intermediate variables that would keep track of
some things that change too often like mouse position, scrolling etc.
Even with these measures in place, the mouse tracking would be quite
noticeably delayed and while not technically a deal breaker, without
that side effect the UI feels much snappier and responsive.

No fps cap is bad because of performance reasons, especially when
scrolling or hovering over a long list of items / widgets.

This current approach is a good bit more performant than no fps cap
but not quite as good as the forced vsync version (because of the lack of
intermediate variables, but we don't want to go back to that) while not
having any of the same side effects.

In the near future I would like to tackle performant lists and trees.
This would eliminate any performance concerns mentioned above.
It should be fairly easy to implement, though there will be some light
reworking of how some things work.
At the end ScrolledBox and TreeView should be able to utilise binary
search to dramatically improve performance and massively raise the limit
to how many items they could hold comfortably. In the far future we
could also think about virtual data in addition to virtual drawing.