Better event handling and drawing that allows for millions of widgets...
Made the necessary changes to be able to use binary search for event
handling and drawing.
At the cost of an extra (reasonable) memory usage we get massively
improved performance. The theoretical performance limit is when binary
search becomes too slow. I have tested it with 10 000 000 buttons and
it can handle that without a sweat. On my machine in a debug build I
get an average of less than 1% CPU usage spamming the heck out of the
scrollbar.

NOTE: There is a little issue with huge numbers of widgets ATM. But
this is not related to the binary search changes.
At a certain point (around 2 400 000 buttons in a test) even though
you scroll to the very end of a ScrolledBox there is a little bit
of empty space where there should be widgets, and it also messes up
the event handling (but only near the end).

In the future we should use the same ideas to massively improve the
TreeView performance.