Change most floats to ints...
This was a change I was thinking about for a long time.
The main issue with floats being used for Size/Point/Rect was that
we dont want subpixel accuracy when it comes to rendering.
In other words, non-whole float values cause issues with rendering
textures, which is why we needed to round in some cases.
It also led to a bug with the ColorPicker that was a hassle to find.
Finally it turned out to be the biggest contributor to the bug that I
mentioned in the ScrolledBox update with binary search.
Changing to ints did not entirely fix that behaviour but it did lessen
it and made the events works accurately again.

The only thing I need to change is to adapt the layout calculations.
Since we cant use fractions anymore we will need to redistribute any
leftover space amongst the children.