Fix buggy behavior with a large number of widgets (tested 18 000 000)...
Changed any floats that should be ints now. Increased precision of
Slider from float to double. This fixed any issues we were having with
a lot of widgets.

Finally removed the out of bounds check for ScrolledBox since its no
longer an problem as it was directly connected to the issue fixed.
However I did find that I needed to account for some rounding,
specifically noticed it because of treeview tab in widget gallery.