Use binary search for fast drawing...
We still have to re-implement tree lines before this can be considered
done.
Regardless the performance is pretty good in row heavy models.
I tested with 1 000 000 rows and everything was smooth as butter.
However in the future we will need to accommodate the uses cases that
utilize a lot of column for which we have a weak point ATM.