Update necessary info for sorting to work correctly...
This changed when we switched how we draw the tree.
We now rely on parent_index on individual nodes to get the next one
from the model, so this needs to be set again. Moreover, we need to
update all the bs_data to reflect the current order of nodes in the model.