Fix lines spanning more than the screen sometimes not drawing...
This happened because we foolishly assumed that the entirety of the
first visible node is on screen.
This resulted in the drawing code not running when it should have.
Now we assume that only the last pixel of the node is visible
which potentially does mean we do a little extra work sometimes
but in practice this has almost no impact on performance.