Add SDL2 and FreeType2 dependencies to src tree to simplify builds...
Change SDL include paths to reflect the SDL2 include folder path in
the SDL2 source.

Removed MSVC libs as we will be building all dependencies as part of
the build process ourselves.

Link SDL2 and FreeType2 to Agro so that executables made with Agro
do not separately need to link those libraries.