Update linking to work on MacOS...
Had the exact same issue before.. unless you specifically add a dir
to the lib search path even an absolute path to the lib will not work.

NOTE: That on MacOS the linker doesnt know how to find libs that are
compiled with debug symbols and suffixed with d.
I can't be fucked to deal with this right now.