Change the logic for tooltip drawing to account for continuous redraws...
Previously once the tooltip would get drawn to the window it would be
turned off and not get drawn again until the user hovered the mouse
over a Widget with a tooltip set.

This becomes problematic when you update a UI many times over a short
period of time since each draw gets rid of the tooltip.
This can easily be seen when you update a ProgressBar.

Now the tooltip only stops getting drawn when the mouse moves.
In the future we can consider doing the same for other kinds of events
but for now this is sufficient.